Remove SVGFormatReader class and distribute its methods
* Anything to do with the DOM goes into a new class, SVGFile. Note that
  this class generally has no dependency on a wiki-based storage
  facility, though it does have a factory method to cover that usage.
* SVGFormatWriter gains getPreferredTranslations, since only it uses
  such a method anyway.
* SVGMessageGroup becomes the owner of all things wiki-related, i.e.
  getOnWikiTranslations().

This adapts TranslateSvg to fit a more test-drive paradigm where
the pretty complicated logic of SVGFile becomes easily testable without
recourse to the creation of endless message groups, which is time and
memory intensive (initial reports suggest 95% time saving).

Also, take advantage of this change to tidy up a number of internal
APIs.

Kudos to Antoine Musso for the inspiration.

Change-Id: Ia2fc5cb8e07787c6039e1c9520a7b6f33284f3a4