Kill standalone PHPUnit and class redefine hack
This cannot easily be ported to extension registration; moreover,
redefining a class like that is an awful hack that should be killed
instantly.

Keeping a standalone PHPUnit config is also unnecessary and brings
nothing good.

Note: mocking TranslateSvgUpload can probably be done by using PHPUnit
mocks without having to create another class, but that's left to do in a
follow-up.

Change-Id: I5b6f5b936b71020c55822cd8dabbf0b1a624953d